requirements.txt
requests
pytest
allure-pytest
curlify
jsonschema==4.18.0